requirements.txt
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attr==0.3.1
attrs==23.2.0
azure-core==1.30.0
azure-storage-blob==12.19.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==5.0.1
boto==2.49.0
boto3==1.34.39
botocore==1.34.39
boxing==0.1.4
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.24
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
executing==2.0.1
expiringdict==1.2.2
fastjsonschema==2.19.1
filelock==3.9.0
fonttools==4.48.1
fqdn==1.5.1
fsspec==2023.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
htmlmin==0.1.12
humansignal-drf-yasg==1.21.9
idna==3.6
ijson==3.2.3
inflection==0.5.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
label-studio==1.11.0
label-studio-converter==0.0.57
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
lockfile==0.12.2
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.6.7
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.26.4
opencv-python==4.9.0.80
ordered-set==4.0.2
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
playwright==1.41.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pyee==11.0.1
Pygments==2.17.2
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
redis==3.5.3
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
semver==2.13.0
Send2Trash==1.8.2
sentry-sdk==1.40.3
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tinycss2==1.2.1
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2023.4
ua-parser==0.18.0
ujson==5.9.0
ultralytics==8.1.11
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.40.0
widgetsnbextension==4.0.10
xmljson==0.2.0
